frontend_service/requirements.txt
gradio
httpx
python-dotenv
pandas # For formatting the output DataFrame if needed